backend/pyproject.toml
[project]
name = "drive-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.24",
    "fastapi>=0.115.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "groq>=1.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "mcp[cli]>=1.25.0",
    "pycryptodome>=3.23.0",
    "pypdf2>=3.0.1",
    "uvicorn>=0.32.0",
]
